- Don't allow to create bitmaps smaller than screen. Not supported yet. - Use RelativeLayout. Makes things working.
git-svn-id: https://svn.aros.org/svn/aros/trunk@40746 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ arch/all-android/bootstrap/app/src/org/aros/bootstrap/AROSBootstrap.java
@@ -129,6 +129,7 @@ public class AROSBootstrap extends Application
 				Bitmap.Height	   = params[4];
 				Bitmap.BytesPerRow = params[5];
 				Bitmap.Pixels      = MapMemory(params[6], params[5] * params[4]);
+//				Bitmap.Pixels	   = MapMemory(0x0A, 16);
 			}
 
 			Server.ReplyCommand(cmd);
@@ -179,5 +180,5 @@ public class AROSBootstrap extends Application
 	
 	// libAROSBootstrap native methods
     private native int Start(String dir, FileDescriptor rfd, FileDescriptor wfd);
-    private native ByteBuffer MapMemory(long addr, long size);
+    private native ByteBuffer MapMemory(int addr, int size);
 }
